requirements.txt
amqp==5.0.6
arrow==0.14.7
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
Django==3.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-filter==2.4.0
django-templated-mail==1.1.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
ics==0.7
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
TatSu==5.6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
